Add edit and delete options to the menu #137

+++ src/systems/deleteSystemDialog.component.tsx
@@ -16,7 +16,7 @@ import { useNavigateToSystem } from './systems.component';
 export interface DeleteSystemDialogProps {
   open: boolean;
   onClose: () => void;
-  system: System;
+  system?: System;
 }
 
 export const DeleteSystemDialog = (props: DeleteSystemDialogProps) => {
@@ -36,31 +36,32 @@ export const DeleteSystemDialog = (props: DeleteSystemDialogProps) => {
   };
 
   const handleDeleteSystem = React.useCallback(() => {
-    deleteSystem(system.id)
-      .then((response) => {
-        onClose();
-        // Navigate to parent as current system no longer exists
-        navigateToSystem(system.parent_id);
-      })
-      .catch((error: AxiosError) => {
-        const response = error.response?.data as ErrorParsing;
-        if (response && error.response?.status === 409) {
-          // Duplicate system
-          setErrorMessage(
-            `${response.detail}, please delete the child systems first`
-          );
-          return;
-        }
-        setErrorMessage('Please refresh and try again');
-      });
-  }, [deleteSystem, navigateToSystem, onClose, system.id, system.parent_id]);
+    if (system)
+      deleteSystem(system.id)
+        .then((response) => {
+          onClose();
+          // Navigate to parent as current system no longer exists
+          navigateToSystem(system.parent_id);
+        })
+        .catch((error: AxiosError) => {
+          const response = error.response?.data as ErrorParsing;
+          if (response && error.response?.status === 409) {
+            // Duplicate system
+            setErrorMessage(
+              `${response.detail}, please delete the child systems first`
+            );
+            return;
+          }
+          setErrorMessage('Please refresh and try again');
+        });
+  }, [deleteSystem, navigateToSystem, onClose, system]);
 
   return (
     <Dialog open={open} onClose={onClose} maxWidth="lg">
       <DialogTitle>Delete System</DialogTitle>
       <DialogContent>
         Are you sure you want to delete{' '}
-        <strong data-testid="delete-system-name">{system.name}</strong>?
+        <strong data-testid="delete-system-name">{system?.name}</strong>?
       </DialogContent>
       <DialogActions>
         <Button onClick={handleClose}>Cancel</Button>
